Refactor lua config files
