+++ lua/alexesba/augroups/file-types.lua
@@ -1,0 +1,11 @@
+vim.api.nvim_create_autocmd(
+  { 'BufRead', 'BufNewFile' }, {
+    pattern = { "*.arb" },
+    command = "setfiletype ruby"
+  })
+
+vim.api.nvim_create_autocmd(
+  { 'BufNewFile', 'BufRead', 'BufEnter' }, {
+    pattern = { "*.jbuilder" },
+    command = "setf jbuilder"
+  });
