+++ lua/alexesba/augroups/init.lua
@@ -1,0 +1,4 @@
+require('alexesba.augroups.column-limit')
+require('alexesba.augroups.file-types')
+require('alexesba.augroups.reload-colorscheme')
+require('alexesba.augroups.reload-vimrc')
