+++ lua/alexesba/augroups/reload-colorscheme.lua
@@ -1,0 +1,9 @@
+local ReloadColorSchemes = vim.api.nvim_create_augroup(
+  "ReloadColorSchemes", { clear = true }
+)
+
+vim.api.nvim_create_autocmd(
+  { "ColorScheme", "VimEnter" }, {
+    group = ReloadColorSchemes,
+    callback = ConfigItalicFonts
+  })
