+++ lua/alexesba/commands/init.lua
@@ -1,0 +1,3 @@
+require('alexesba.commands.formatters')
+require('alexesba.commands.telescope')
+require('alexesba.commands.vim-table-mode')
