+++ lua/alexesba/commands/telescope.lua
@@ -1,0 +1,9 @@
+local telescope = prequire('telescope')
+
+if (telescope) then
+  vim.api.nvim_create_user_command(
+    'ColorScheme',
+    require('telescope.builtin').colorscheme,
+    { desc = 'Reload VIMRC file' }
+  )
+end
