+++ lua/alexesba/init.lua
@@ -1,10 +1,10 @@
-require('alexesba.nvim')
-require('alexesba.packerBootstrap')
-require('alexesba.packer')
+require('alexesba.base')
+require('alexesba.packer.bootstrap')
+require('alexesba.packer.plugins')
 require('alexesba.functions')
 require('alexesba.commands')
 require('alexesba.mappings')
 require('alexesba.augroups')
 require('alexesba.statusLine')
-require('alexesba.lsp')
-require('alexesba.autoreload')
+require('alexesba.linters')
+require('alexesba.utils.autoreload')
