+++ lua/alexesba/linters/remark_ls.lua
@@ -1,3 +1,3 @@
-local lspServer = require('utils.lspServer')
+local lspServer = require('alexesba.utils.lspServer')
 
 lspServer.configure('remark_ls')
