+++ lua/alexesba/linters/solargraph.lua
@@ -1,4 +1,4 @@
-local lspServer = require('utils.lspServer')
+local lspServer = require('alexesba.utils.lspServer')
 
 lspServer.configure('solargraph', {
   diagnostic = true,
