+++ lua/alexesba/mappings/bundle.lua
@@ -1,0 +1,3 @@
+local map = require('alexesba.utils.map').map
+
+map("n", "<leader>bo", "':Bopen '. expand('<cword>') .'<cr>'", { noremap = true, expr = true })
