+++ lua/alexesba/mappings/markdown-preview.lua
@@ -1,0 +1,2 @@
+local map = require('alexesba.utils.map').map
+map('n', '<leader>mp', "<cmd>MarkdownPreviewToggle<cr>", { silent = true, noremap = true })
