+++ lua/alexesba/mappings/nvim-code-action-menu.lua
@@ -1,0 +1,4 @@
+local map = require('alexesba.utils.map').map
+-- auto complete menu
+map("n", "<leader>ac", ":CodeActionMenu<CR>", { silent = true })
+
