+++ lua/alexesba/mappings/nvim-tree.lua
@@ -1,0 +1,8 @@
+local map = require('alexesba.utils.map').map
+map(
+  "n",
+  "<C-p>",
+  ":NvimTreeFindFileToggle!<CR>", {
+    silent = true,
+    noremap = true
+  });
